requirements.txt
google-cloud-bigquery==3.17.2
google-cloud-logging==3.9.0
google-cloud-pubsub==2.31.1
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
pyyaml==6.0.1
pyarrow==16.1.0
jupyterlab==4.1.5
db-dtypes==1.2.0
